% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_mrmode.R
\name{mv_mrmode}
\alias{mv_mrmode}
\title{MVMR Mode: A plurality valid sensitivity analysis for multivariable MR}
\usage{
mv_mrmode(
  Bout,
  Bexp,
  SEout,
  SEexp,
  Psi = 0,
  CIMin = NA,
  CIMax = NA,
  CIStep = 0.001,
  alpha = 0.05,
  residual = "IVW",
  Mode = "CM",
  weighting = "weighted",
  stderror = "simple",
  phi = 1,
  distribution = "normal",
  iterations = 10000
)
}
\arguments{
\item{Bout}{SNP-outcome associations.}

\item{Bexp}{SNP-exposure associations.}

\item{SEout}{Standard error of the SNP-outcome associations}

\item{SEexp}{Standard error of the SNP-exposure associations.}

\item{Psi}{tuning parameter for the contamination mixture (CM) method (see the original method's documentation for more details)}

\item{CIMin}{The lower bound on the 95\% CI to search for when using CM}

\item{CIMax}{The upper bound of the 95\% CI to search for when using CM}

\item{CIStep}{The space between each step in CM (see the original method's documentation for more details)}

\item{alpha}{The p-value threshold}

\item{residual}{the method used to create residuals. Can be either "IVW", "Egger", or  "Robust". N.B. the method has only been validated for the defult (IVW)}

\item{Mode}{The type of plurality valid  estimator to use. Can be "CM" for contamination mixture, or "MBE" for a traditional mode based estimator. n.b. the default (CM) performed best in the validation}

\item{weighting}{For the MBE, if to use a weighed MBE or weighted.}

\item{stderror}{For MBE, the type of standard error. Please see MendelianRandomization R package documentation for more detail}

\item{phi}{tuning parameter for the MBE. Please see MendelianRandomization R package documentation for more detail
@param distribution distribution assumed for the MBE. Please see MendelianRandomization R package documentation for more detail
@param iterations the number of iterations to use for the SE of the MBE}
}
\description{
This function implements a plurality valid (e.g. mode based) estimator for MVMR
}
\keyword{2SMR,}
\keyword{MR}
\keyword{MVMR,}
\keyword{Mendelian}
\keyword{Multivaribile}
\keyword{Randomization,}
\keyword{Two-sample}
